requirements.txt
asttokens==2.1.0
backcall==0.2.0
click==8.1.3
configobj==5.0.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
Flask-Markdown==0.3
Flask-WTF==1.0.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
stack-data==0.6.0
traitlets==5.5.0
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
# Data processing & ingestion
luaparser
numpy
h5py
openpyxl
pandas
psycopg2-binary
requests
sqlalchemy
xlrd

# ML & plotting (SC ↔ irradiation equivalence analysis)
scikit-learn>=1.3
matplotlib>=3.7
